// src/utils/types/dashboard.ts
export interface DashboardQueryParams {
    _id?: string;
    page_size?: string;
    limit?: string;
    status?: string;
    sortby?: string;
    sortorder?: string;
    keyword?: string;
    page?: string;
    pageReference?: string;
    countryId?: string;
    fromDate?: string;
    endDate?: string;
}